example renaming and cleanup
