+++ apps/docs/src/examples/extras/instanced-sprite/App.svelte
@@ -2,6 +2,7 @@
   import { Canvas, T } from '@threlte/core'
   import Scene from './Scene.svelte'
   import Settings from './Settings.svelte'
+  import { OrbitControls, PerfMonitor } from '@threlte/extras'
 
   let billboarding = false
   let fps = 10
@@ -9,6 +10,15 @@
 
 <div>
   <Canvas>
+    <PerfMonitor />
+
+    <T.PerspectiveCamera
+      makeDefault
+      position.z={14}
+      position.y={6}
+    >
+      <OrbitControls />
+    </T.PerspectiveCamera>
     <Scene
       {billboarding}
       {fps}
