+++ apps/docs/src/examples/extras/instanced-sprite/SpriteAtlasTrees.svelte
