+++ apps/docs/src/examples/extras/instanced-sprite/SpriteDudes.svelte
