+++ apps/docs/src/examples/extras/instanced-sprite/SpriteFlyersTyped.svelte
