+++ apps/docs/src/examples/extras/instanced-sprite/SpriteGoblin.svelte
