+++ packages/extras/package.json
@@ -70,6 +70,7 @@
     "types": "./dist/index.d.ts"
   },
   "dependencies": {
+    "troika-three-text": "^0.49.0",
     "@threejs-kit/instanced-sprite-mesh": "^2.4.3",
     "three-mesh-bvh": "^0.7.1",
     "three-perf": "^1.0.10"
